+++ packages/carbon-graphs/dev/examples/controls/colors.js
@@ -1,4 +1,4 @@
-import * as d3 from 'd3';
+import * as d3 from '../../../src/js/d3Modules';
 import Carbon from '../../../src/js/carbon';
 
 import { vector } from '../helpers';