+++ packages/carbon-graphs/src/js/controls/Bar/helpers/goalLineHelpers.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import * as d3 from 'd3';
+import * as d3 from '../../../d3Modules';
 import { parseTypedValue } from '../../../core/BaseConfig';
 import { isValidAxisType } from '../../../helpers/axis';
 import constants from '../../../helpers/constants';