+++ packages/carbon-graphs/src/js/controls/Bubble/helpers/colorGradient.js
@@ -1,4 +1,4 @@
-import * as d3 from 'd3';
+import * as d3 from '../../../d3Modules';
 import constants, { BUBBLE } from '../../../helpers/constants';
 import utils from '../../../helpers/utils';