+++ packages/carbon-graphs/src/js/controls/Graph/Graph.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import * as d3 from 'd3';
+import * as d3 from '../../d3Modules';
 import Construct from '../../core/Construct';
 import { contentHandler } from '../../helpers/constructUtils';
 import {