+++ packages/carbon-graphs/src/js/controls/PairedResult/helpers/selectionIndicatorHelpers.js
@@ -1,4 +1,4 @@
-import * as d3 from 'd3';
+import * as d3 from '../../../d3Modules';
 import Shape, { getDefaultSVGProps } from '../../../core/Shape';
 import constants, { SHAPES } from '../../../helpers/constants';
 import styles from '../../../helpers/styles';