+++ packages/carbon-graphs/src/js/controls/Timeline/helpers/translateHelpers.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import * as d3 from 'd3';
+import * as d3 from '../../../d3Modules';
 import { getRotationForAxis, getYAxisHeight } from '../../../helpers/axis';
 import constants from '../../../helpers/constants';
 import styles from '../../../helpers/styles';