+++ packages/carbon-graphs/src/js/core/Shape/helpers.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import * as d3 from 'd3';
+import * as d3 from '../../d3Modules';
 import constants from '../../helpers/constants';
 import errors from '../../helpers/errors';
 import styles from '../../helpers/styles';