+++ packages/carbon-graphs/src/js/helpers/legend.js
@@ -3,7 +3,7 @@
  * @alias module:legend
  */
 
-import * as d3 from 'd3';
+import * as d3 from '../d3Modules';
 import {
   d3RemoveElement,
   getColorForTarget,