+++ packages/carbon-graphs/src/js/helpers/transformUtils.js
@@ -1,4 +1,4 @@
-import * as d3 from 'd3';
+import * as d3 from '../d3Modules';
 
 /**
  * Returns the current d3 transform attribute object from context object