+++ packages/carbon-graphs/tests/unit/controls/Timeline/TimelineUnload-spec.js
@@ -1,6 +1,7 @@
 'use strict';
 
-import * as d3 from 'd3';
+import * as d3 from '../../../../src/js/d3Modules';
+
 import Timeline from '../../../../src/js/controls/Timeline';
 import { getXAxisYPosition } from '../../../../src/js/controls/Timeline/helpers/creationHelpers';
 import { getYAxisHeight } from '../../../../src/js/helpers/axis';