+++ packages/carbon-graphs/tests/unit/helpers/commonHelpers.js
@@ -1,4 +1,4 @@
-import * as d3 from 'd3';
+import * as d3 from '../../../src/js/d3Modules';
 /**
  * @type {number} Delay duration
  */