+++ packages/terra-graphs-docs/src/terra-dev-site/graph/example/Graphs/CombinationGraph/BasicCombinationGraphs/LineWithSingleBubbleData.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
-import CombinationGraph from '@cerner/terra-graphs/lib/components/Combination/CombinationGraph';
+import CombinationGraph from '@cerner/terra-graphs-docs/lib/terra-graphs-src/components/Combination/CombinationGraph';
 import '@cerner/terra-graphs-docs/lib/terra-dev-site/ExampleGraphContainer/ExampleGraphContainer.module.scss';
 import getCombinationGraphConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Combination/combinationDefault';
 import lineDataset from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Combination/lineDataset1';