+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated d3 dependencies to only pull in `d3-selection` as the other modules are not needed.
+
 ## 1.4.0 - (March 29, 2023)
 
 * Changed